fix: patch __proto__ if not extending element
