+++ src/package/polyfill/constructInstance.ts
@@ -1,3 +1,4 @@
+import { patch } from './patch';
 export const BLACKLISTED_PROTOTYPE_PATCH_METHODS = [
     'constructor',
     'connectedCallback',
@@ -6,67 +7,35 @@ export const BLACKLISTED_PROTOTYPE_PATCH_METHODS = [
     'attributeChangedCallback'
 ];
 
+export const BLACKLISTED_STATIC_PATCH_METHODS = ['name', 'prototype', 'length'];
+
 export function constructInstance(mostRecentImpl: any, args: any, newTarget: any) {
     // Constructed instance partly points to outdated impl details.
     // This patch loop makes sure that the hook methods aren't overridden,
     // the constructor stays intact but methods, getters, setters and fields
     // are updated according to the most recent implementation:
 
-    const prototypePropertyNames = Object.getOwnPropertyNames(mostRecentImpl.prototype);
-
-    const whitelistedPrototypePropertyNames = prototypePropertyNames.filter(
-        (propertyName: string) => {
-            return BLACKLISTED_PROTOTYPE_PATCH_METHODS.indexOf(propertyName) === -1;
-        }
-    );
-
-    for (let i = 0; i < whitelistedPrototypePropertyNames.length; i++) {
-        const propertyDescriptor = Object.getOwnPropertyDescriptor(
-            mostRecentImpl.prototype,
-            whitelistedPrototypePropertyNames[i]
-        );
-
-        if (propertyDescriptor) {
-            if (propertyDescriptor.configurable) {
-                Object.defineProperty(
-                    newTarget.prototype,
-                    whitelistedPrototypePropertyNames[i],
-                    propertyDescriptor
-                );
-            } else {
-                console.warn(
-                    '[custom-element-hmr-polyfill]',
-                    `${whitelistedPrototypePropertyNames[i]} is not configurable, skipping`
-                );
-            }
-        }
+    // PROTO check
+    let check: any = window[mostRecentImpl.__proto__.name];
+    if (check) {
+        check = (window[mostRecentImpl.__proto__.name] as any).prototype instanceof Element;
     }
 
-    // here we will update static variables/methods
-
-    const ownPropertyNames = Object.getOwnPropertyNames(mostRecentImpl);
-
-    const whitelistedPropertyNames = ownPropertyNames.filter((propertyName: string) => {
-        return ['name', 'prototype', 'length'].indexOf(propertyName) === -1;
-    });
-
-    for (let i = 0; i < whitelistedPropertyNames.length; i++) {
-        const propertyDescriptor = Object.getOwnPropertyDescriptor(
-            mostRecentImpl,
-            whitelistedPropertyNames[i]
+    if (!check) {
+        patch(
+            mostRecentImpl.__proto__.prototype,
+            newTarget.prototype,
+            BLACKLISTED_PROTOTYPE_PATCH_METHODS
         );
-
-        if (propertyDescriptor) {
-            if (propertyDescriptor.configurable) {
-                Object.defineProperty(newTarget, whitelistedPropertyNames[i], propertyDescriptor);
-            } else {
-                console.warn(
-                    '[custom-element-hmr-polyfill]',
-                    `${whitelistedPropertyNames[i]} is not configurable, skipping`
-                );
-            }
-        }
+        // here we will update static variables/methods of "__proto__"
+        patch(mostRecentImpl.__proto__, newTarget, BLACKLISTED_STATIC_PATCH_METHODS);
     }
+
+    // PROTOTYPE
+    patch(mostRecentImpl.prototype, newTarget.prototype, BLACKLISTED_PROTOTYPE_PATCH_METHODS);
+
+    // here we will update static variables/methods of class
+    patch(mostRecentImpl, mostRecentImpl, BLACKLISTED_STATIC_PATCH_METHODS);
 
     const customElementInstance = Reflect.construct(mostRecentImpl, args, newTarget);
